+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 ### Changed
 - Dependencies - Updated CSharpIsNullAnalyzer to 0.2.19
 - Dependencies - Updated Meziantou.Analyzer to 3.0.177
+- Dependencies - Updated Roslynator.Analyzers to 5.0.0
 ### Deprecated
 ### Removed
 ### Deployment Changes